# src/main/java/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "textSaver"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "tags": [
    {
      "name": "База данных с людьми",
      "description": "В базе содержатся имена людей"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/person": {
      "post": {
        "tags": [
          "База данных с людьми"
        ],
        "summary": "Добавляет человека в базу данных",
        "operationId": "addPerson",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Имя человека",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Person"
              }
            }
          }
        }
      }
    },
    "/api/person/{id}": {
      "get": {
        "tags": [
          "База данных с людьми"
        ],
        "summary": "Получить имя человека",
        "operationId": "getPersonById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID человека",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Person"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "База данных с людьми"
        ],
        "summary": "Удаляет человека из базы",
        "operationId": "PersonDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID человека",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Person": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}